marketing/requirements.txt
anthropic>=0.40.0
tweepy>=4.14.0
praw>=7.7.0
python-telegram-bot>=21.0
discord.py>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
